+++ flopp-kernel/src/main/java/com/github/kjetilv/flopp/kernel/bits/MemorySegments.java
@@ -29,7 +29,7 @@ public final class MemorySegments {
 
         long offset();
 
-        int length();
+        long length();
 
         default byte[] asBytes() {
             return toBytes(this);
